updated the deleted functionality

+++ src/Dashboard/Todo.jsx
@@ -15,20 +15,34 @@ const Todo = () => {
     }, [])
 
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:5000/allTasksData/${id}`)
-            .then((res) => {
-                if (res.data.deletedCount > 0) {
-                    Swal.fire({
-                        position: "top-end",
-                        icon: "success",
-                        title: "task deleted Successfully",
-                        showConfirmButton: false,
-                        timer: 1500
-                    });
-                    const remaining = task.filter(remain => remain._id !== id);
-                    setTask(remaining);
-                }
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:5000/allTasksData/${id}`)
+                    .then((res) => {
+                        if (res.data.deletedCount > 0) {
+                            Swal.fire({
+                                position: "top-end",
+                                icon: "success",
+                                title: "task deleted Successfully",
+                                showConfirmButton: false,
+                                timer: 1500
+                            });
+                            const remaining = task.filter(remain => remain._id !== id);
+                            setTask(remaining);
+                        }
+                    })
+            }
+        });
+
+
     }
 
     return (
